:scissors: keep only lib files
