--- src/App.css
@@ -1,33 +1,0 @@
-.App {
-  height: 100vh;
-  display: flex;
-}
-
-.container {
-  width: 1170px;
-  margin: 0px auto;
-}
-
-.filter-group {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin: 20px 0;
-}
-
-.filter-group div {
-  display: flex;
-  align-items: center;
-}
-
-input,
-select,
-input:focus,
-select:focus {
-  display: inline-block;
-  padding: 5px 10px;
-  font-size: 14px;
-  border: 0;
-  border-bottom: 1px solid #ccc;
-  outline: none;
-}
