+++ src/index.css
@@ -1,13 +1,45 @@
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen", "Ubuntu", "Cantarell", "Fira Sans",
+    "Droid Sans", "Helvetica Neue", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
+}
+
+.App {
+  height: 100vh;
+  display: flex;
+}
+
+.container {
+  width: 1170px;
+  margin: 0px auto;
+}
+
+.filter-group {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 20px 0;
+}
+
+.filter-group div {
+  display: flex;
+  align-items: center;
+}
+
+input,
+select,
+input:focus,
+select:focus {
+  display: inline-block;
+  padding: 5px 10px;
+  font-size: 14px;
+  border: 0;
+  border-bottom: 1px solid #ccc;
+  outline: none;
 }
